+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - PHP 7.4 compatibility issue: deprecated parameter order of implode()
 - PHP 7.4 test coverage
+- Solarium\Component\Result\Stats\Result getters might return null
 
 
 ## [5.1.4]